Revert "Merge remote-tracking branch 'origin/develop' into secure-central-docker"
This reverts commit 292b157bb071e7c18771da22444ae3ae58315cb8, reversing
changes made to af6452a908b943ddfac64d1797320e8cb04d04c9.
