+++ .github/workflows/build-dependency-cache.yml
@@ -1,14 +1,8 @@
 ---
 # Creates a cache of the env directory
-# Documentation for the syntax of this file is located
-# https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-syntax-for-github-actions
-
-# The workflow name will show up in the action tab on github during execution
-# https://github.com/VOLTTRON/volttron/actions (or if you are pushing to your own fork change the user)
 
 name: cache virtual env
 
-# Check the cache when a push happens to the repository.
 on: [push]
 
 jobs:
@@ -16,6 +10,8 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
+        #ubuntu-16.04
+        #os: [ubuntu-18.04, ubuntu-20.04]
         os: [ubuntu-16.04, ubuntu-18.04, ubuntu-20.04]
         python-version: [3.6, 3.7, 3.8]
     steps:
@@ -53,7 +49,7 @@ jobs:
         if: steps.check_files.outputs.files_exists != 'true'
         run: |
           pip install wheel
-          python bootstrap.py --all --force
+          python bootstrap.py --all
 
 # Only works on default branch of the target repo
 #       - name: Repository Dispatch
