+++ volttron/platform/__init__.py
@@ -41,14 +41,12 @@
 
 import logging
 import os
-import traceback
-
 import psutil
 import sys
 from configparser import ConfigParser
+from ..utils.frozendict import FrozenDict
 from urllib.parse import urlparse
 
-from ..utils.frozendict import FrozenDict
 __version__ = '8.0-rc'
 
 _log = logging.getLogger(__name__)
@@ -93,42 +91,14 @@ def get_config_path() -> str:
     return os.path.join(get_home(), "config")
 
 
-def get_address(verify_listening=False):
+def get_address():
     """Return the VIP address of the platform
-    If the VOLTTRON_VIP_ADDR environment variable is set, it is used to connect to.
+    If the VOLTTRON_VIP_ADDR environment variable is set, it used.
     Otherwise, it is derived from get_home()."""
     address = os.environ.get('VOLTTRON_VIP_ADDR')
     if not address:
-        # Connect via virtual unix socket if linux platform (mac doesn't have @ in it)
         abstract = '@' if sys.platform.startswith('linux') else ''
         address = 'ipc://%s%s/run/vip.socket' % (abstract, get_home())
-
-    import zmq.green as zmqgreen
-    import zmq
-    # The following block checks to make sure that we can
-    # connect to the zmq based upon the ipc address.
-    #
-    # The zmq.sock.bind() will raise an error because the
-    # address is already bound (therefore volttron is running there)
-    sock = None
-    try:
-        # TODO: We should not just do the connection test when verfiy_listening is True but always
-        # Though we leave this here because we have backward compatible unit tests that require
-        # the get_address to not have somethiing bound to the address.
-        if verify_listening:
-            ctx = zmqgreen.Context.instance()
-            sock = ctx.socket(zmq.PUB)  # or SUB - does not make any difference
-            sock.bind(address)
-            raise ValueError("Unable to connect to vip address "
-                             f"make sure VOLTTRON_HOME: {get_home()} "
-                             "is set properly")
-    except zmq.error.ZMQError as e:
-         print(f"Zmq error was {e}\n{traceback.format_exc()}")
-    finally:
-        try:
-            sock.close()
-        except AttributeError as e:  # Raised when sock is None type
-            pass
 
     return address
